# src/core/Snakk.Sdk/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Snakk.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost:17100/"
    }
  ],
  "paths": {
...
    "/admin/permissions/roles/{roleId}": {
      "get": {
        "tags": [
          "Admin Permissions"
        ],
        "operationId": "GetRolePermissions",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRolePermissionsResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Permissions"
        ],
        "operationId": "AssignPermissionToRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignPermissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPermissionActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/permissions/roles/{roleId}/{permissionId}": {
      "delete": {
        "tags": [
          "Admin Permissions"
        ],
        "operationId": "RevokePermissionFromRole",
        "parameters": [
          {
            "name": "roleId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "permissionId",
            "in": "path",
            "required": true,
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPermissionActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/temporary-roles": {
      "get": {
        "tags": [
          "Admin Temporary Roles"
        ],
        "operationId": "GetActiveTemporaryRoles",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTemporaryRolesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin Temporary Roles"
        ],
        "operationId": "GrantTemporaryRole",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantTemporaryRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTemporaryRoleGrantResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/temporary-roles/{elevationId}": {
      "delete": {
        "tags": [
          "Admin Temporary Roles"
        ],
        "operationId": "RevokeTemporaryRole",
        "parameters": [
          {
            "name": "elevationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevokeTemporaryRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminTemporaryRoleRevokeResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminPermissionActionResponse": {
        "required": [
          "message",
          "roleId",
          "permissionId"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "roleId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "permissionId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "AdminRolePermissionsResponse": {
        "required": [
          "roleId",
          "permissions",
          "total"
        ],
        "type": "object",
        "properties": {
          "roleId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "permissions": {
            "type": "array"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "AdminTemporaryRoleGrantResponse": {
        "required": [
          "message",
          "elevation"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "elevation": { }
        }
      },
      "AdminTemporaryRoleRevokeResponse": {
        "required": [
          "message",
          "elevationId"
        ],
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "elevationId": {
            "type": "string"
          }
        }
      },
      "AdminTemporaryRolesResponse": {
        "required": [
          "elevations",
          "total"
        ],
        "type": "object",
        "properties": {
          "elevations": {
            "type": "array"
          },
          "total": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "AssignPermissionRequest": {
        "required": [
          "permissionId"
        ],
        "type": "object",
        "properties": {
          "permissionId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
...
      "GrantTemporaryRoleRequest": {
        "required": [
          "userId",
          "roleType",
          "scope",
          "scopeId",
          "expiresAt",
          "reason"
        ],
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "roleType": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "scopeId": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "RevokeTemporaryRoleRequest": {
        "required": [
          "reason"
        ],
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}